+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3ClassicAirDecompileTest.java
@@ -1507,6 +1507,34 @@ public class ActionScript3ClassicAirDecompileTest extends ActionScript3Decompile
     }
 
     @Test
+    public void testSwitchContinue() {
+        decompileMethod("classic_air", "testSwitchContinue", "var i:int = 0;\r\n"
+                + "var r:int = Math.random() % 10;\r\n"
+                + "if(r > 5)\r\n"
+                + "{\r\n"
+                + "for(i = 0; i < 10; i++)\r\n"
+                + "{\r\n"
+                + "switch(i)\r\n"
+                + "{\r\n"
+                + "case 0:\r\n"
+                + "trace(\"hello\");\r\n"
+                + "break;\r\n"
+                + "case 1:\r\n"
+                + "trace(\"hi\");\r\n"
+                + "break;\r\n"
+                + "case 2:\r\n"
+                + "trace(\"howdy\");\r\n"
+                + "break;\r\n"
+                + "default:\r\n"
+                + "continue;\r\n"
+                + "}\r\n"
+                + "trace(\"message shown\");\r\n"
+                + "}\r\n"
+                + "}\r\n",
+                 false);
+    }
+
+    @Test
     public void testSwitchDefault() {
         decompileMethod("classic_air", "testSwitchDefault", "var a:int = 5;\r\n"
                 + "switch(a)\r\n"